no-method-argument (E0211)
Changelog:

- Model class deleted
- src.model.db moved to src.model
- Parse class from `src.model.__parser__` deleted
- Imports on `src` starts with `src.`
